backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
requests
python-dotenv
sqlalchemy
psycopg2-binary
apscheduler
